Improved Surround bypassing
